% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.elem.R
\name{extract.elem}
\alias{extract.elem}
\title{Extract a Model Fit Element}
\usage{
extract.elem(.data, .formula, .fun, elem, mod.sum = FALSE, ...)
}
\arguments{
\item{.data}{Data frame.}

\item{.formula}{Formula to use with .fun.}

\item{.fun}{Model function such as: "lm".}

\item{elem}{Element to extract from fit object, such as: "coefficients".}

\item{mod.sum}{If FALSE (default), summary() is not run on model fit.}

\item{...}{Additional options for .fun such as: family = "inverse.gaussian".}
}
\value{
Data frame combining an element of fit results for many models.
}
\description{
Extracts an element from a model fit of many models as a data frame.
}
\examples{
library(datasets)
data(mtcars)

formulas <- c('mpg ~ cyl', 'mpg ~ cyl + disp', 'mpg ~ cyl + disp + hp')

set.seed(1)
weights <- c(abs(rnorm(nrow(mtcars))))

df <- extract.elem(.data = mtcars, .formula = formulas,
                   .fun = 'lm', elem = 'coefficients')

df.sum <- extract.elem(.data = mtcars, .formula = formulas, .fun = "lm",
                       elem = "coefficients", mod.sum = TRUE)

df.residuals <- extract.elem(.data = mtcars, .formula = formulas,
                             .fun = 'lm', elem = 'residuals')

df.weighted <- extract.elem(.data = mtcars, .formula = formulas,
                            .fun = 'lm', elem = 'coefficients',
                            weights = weights)

df.inv.gauss <- extract.elem(.data = mtcars, .formula = formulas,
                             .fun = 'glm', elem = 'coefficients',
                             family = "inverse.gaussian")

}
