% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elem.to.wide.R
\name{elem.to.wide}
\alias{elem.to.wide}
\title{Reshape a Model Fit Element}
\usage{
elem.to.wide(.data, idvar = "formula")
}
\arguments{
\item{.data}{Data frame with "variable", "value", and "formula" columns.}

\item{idvar}{Names of one or more grouping variables, passed to reshape().}
}
\value{
Data frame in "Wide" format.
}
\description{
Reshapes as extracted model fit element data frame to "wide" format.
}
\examples{
library(datasets)
data(mtcars)
formulas <- c('mpg ~ cyl', 'mpg ~ cyl + disp', 'mpg ~ cyl + disp + hp')
df <- extract.elem(.data = mtcars, .formula = formulas,
                   .fun = 'lm', elem = 'coefficients')
df.wide <- elem.to.wide(df)
df.wide

}
